requirements.txt
# Core web framework
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2

# Database
sqlalchemy==2.0.41
psycopg2-binary>=2.9.11
sqlmodel>=0.0.24

# Web scraping
beautifulsoup4==4.13.4
requests==2.32.4
selectolax>=0.3.21
soupsieve==2.7

# AWS services
boto3>=1.40.62

# AI/ML
google-genai>=1.43.0

# Logging
loguru>=0.7.3

# Environment variables
python-dotenv==1.1.1

# Data validation (required by FastAPI)
pydantic==2.11.7
pydantic_core==2.33.2

# Required by various packages
typing_extensions==4.14.1
annotated-types==0.7.0

# HTTP and networking (required by requests/fastapi)
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
anyio==4.9.0
sniffio==1.3.1
idna==3.10
certifi==2025.6.15
charset-normalizer==3.4.2
urllib3==2.5.0

# Other required dependencies
click==8.2.1